create Location Ok

+++ GameTrip/GameTrip.Domain/Entities/Location.cs
@@ -5,8 +5,8 @@ public class Location
     public Guid IdLocation { get; set; }
     public string? Name { get; set; }
     public string? Description { get; set; }
-    public double Latitude { get; set; }
-    public double Longitude { get; set; }
+    public decimal Latitude { get; set; }
+    public decimal Longitude { get; set; }
 
     public ICollection<Picture>? Pictures { get; set; }
     public ICollection<Game>? Games { get; set; }
